requirements.txt
asgiref==3.5.2
Django==4.1.2
django-debug-toolbar==3.7.0
gunicorn==20.1.0
Pillow==9.2.0
sqlparse==0.4.3
tzdata==2022.4
whitenoise==6.2.0

djangorestframework~=3.14.0

djangorestframework-simplejwt==5.2.2